Created channel's archive manager
It is disabled by default for all channels. When you enable it for some
channels, it will parse the epg data and populate the database.
In the case of a virtual channel, a queue will be created from all
elements of the epg, than queue will be processed one by one and
collect additional data into the database.
Keep in mind that the more channels you have enabled - the larger
database file will be and the longer the epg file update process will take.